refactor: validator 위치 변경
