+++ src/test/java/racingcar/domain/CarsTest.java
@@ -8,7 +8,7 @@ import java.util.StringJoiner;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.CsvSource;
-import racingcar.view.InputView;
+import racingcar.util.Validator;
 
 public class CarsTest {
     private Cars cars;
@@ -21,7 +21,7 @@ public class CarsTest {
     @ParameterizedTest
     @CsvSource(value = {"글자수제한", "자동차1,자동차20"}, delimiter = '/')
     public void 자동차_이름_부여(String namesFromUser) {
-        String[] names = InputView.validateNames(namesFromUser);
+        String[] names = Validator.validateNames(namesFromUser);
         cars.addCarsFromNames(names);
         List<Car> list = cars.getCars();
