+++ src/test/java/racingcar/view/InputViewTest.java
@@ -7,6 +7,7 @@ import org.junit.jupiter.params.provider.CsvSource;
 import org.junit.jupiter.params.provider.EmptySource;
 import org.junit.jupiter.params.provider.NullAndEmptySource;
 import org.junit.jupiter.params.provider.NullSource;
+import racingcar.util.Validator;
 
 public class InputViewTest {
 
@@ -16,7 +17,7 @@ public class InputViewTest {
     @NullAndEmptySource
     @CsvSource({"-1", "-99", "string", "s", "가", "1 0", "0 8", "1*9", "1+1"})
     void 시도횟수_예외처리(String tryCountFromUser) {
-        assertThatThrownBy(() -> InputView.validateTryCount(tryCountFromUser))
+        assertThatThrownBy(() -> Validator.validateTryCount(tryCountFromUser))
                 .isInstanceOf(IllegalArgumentException.class);
     }
 
@@ -26,7 +27,7 @@ public class InputViewTest {
     @NullAndEmptySource
     @CsvSource(value = {"자동차.자동차2345", "자동차1,racing"}, delimiter = '/')
     void 자동차이름_예외처리(String namesFromUser) {
-        assertThatThrownBy(() -> InputView.validateNames(namesFromUser))
+        assertThatThrownBy(() -> Validator.validateNames(namesFromUser))
                 .isInstanceOf(IllegalArgumentException.class);
     }
 }
